Utilize constructor method to deep clone on spawn task.
Without this, in an electron environment there is
a risk that the entire object is not cloned properly.

Sadly I do not know exactly how to perform this operation
within TypeScripts linting rules, so until then I've marked
this with an ignore.

Basically calling an `Object.create(this)` doesn't properly
initialize all the prototype functions/members. This leads
to an Illegal Instruction error when we try to call `this.signal`,
which doesn't exist and the runtime fails to hold the context
of `this` correctly.

This may need to be a pattern implemented elsewhere in the package,
though this was the only place I currently hit the issue. Perhaps
there is a better, more typescript friendly way, to perform this.